Fixed a bug for the Shap explanation with multiple sample for the detection. Modified the dependence plot for the detection

+++ Neurlux/neurlux_classification.py
@@ -164,7 +164,7 @@ if __name__ == '__main__':
 
             # Subset
             x=[]
-            x_tokens = np.zeros(shape=(N_SAMPLES_EXP * len(classes), x_ts_tokens.shape[1]))
+            x_tokens = np.zeros(shape=(N_SAMPLES_EXP * len(classes), x_ts_tokens.shape[1]),dtype=int)
             y = np.zeros(shape=(N_SAMPLES_EXP * len(classes)), dtype=int)
             for i in range(len(classes)):
                 idx = (y_ts == i).to_numpy()
@@ -217,7 +217,7 @@ if __name__ == '__main__':
         elif EXP_MODE == "multi":
 
             # Subset
-            sample_tokens = np.zeros(shape=(N_SAMPLES_EXP * len(classes), x_ts_tokens.shape[1]))
+            sample_tokens = np.zeros(shape=(N_SAMPLES_EXP * len(classes), x_ts_tokens.shape[1]),dtype=int)
             idx_true = np.zeros(shape=(N_SAMPLES_EXP * len(classes)),dtype=int)
             for i in range(len(classes)):
                 idx = (y_ts == i).to_numpy()
